🚚 rename: 파일명 변경 및 불필요 파일 삭제
- 파일명 변경 MentoringSidebar -> MentoringInfoWithApply
- 파일 삭제: ProfileImage.tsx
