--- src/components/mentoringDetail/MentorProfileImage.tsx
@@ -1,15 +1,0 @@
-const MentorProfileImage = () => {
-	return (
-		<div>
-			<div className="h-[10rem] bg-black-500">
-				<img
-					src="src/assets/image/sample.jpg"
-					alt="sample"
-					className="w-full h-full object-cover"
-				/>
-			</div>
-		</div>
-	);
-};
-
-export default MentorProfileImage;
